Finalize data after analysis

Function that is used internally to change the data back to their original format after conducting an analysis, it being either binning (bin()) or using the Kalman filter (kalman_filter()). The finalization consists of renaming the columns to the user-defined column names instead of the package- required ones and deleting the grouping variable if it was not originally present in the data.

Usage

finalize(data, cols = NULL, .by = NULL)

Arguments

data

Dataframe that contains information on location (x- and y-coordinates) and the time at which the measurement was taken. It is assumed that this information is contained within the columns "x", "y", and "time" respectively.

cols

Named vector or named list containing the mapping of the original column names to the internal ones used within the package. Defaults to NULL, therefore assuming the structure explained in data.

.by

String denoting whether the moving window should be taken with respect to a given grouping variable. Defaults to NULL.

Value

Adjusted data.frame

See also

Examples

# Generate data for illustration purposes
#
# Note that for this to work, I need to define the package-required column 
# names
data <- data.frame(
  x = rnorm(100),
  y = rnorm(100),
  time = rep(1:50, times = 2) / 10,
  id = rep(1:2, each = 50)
)

# Prepare the data for analysis
finalize(
  data,
  cols = c(
    "time" = "seconds",
    "x" = "X",
    "y" = "Y"
  ),
  .by = "tag"
)
#>     seconds            X           Y
#> 1       0.1 -0.506286298 -0.52079637
#> 2       0.2 -1.109968853 -0.63901860
#> 3       0.3 -0.948705723 -0.63589414
#> 4       0.4  0.476843757  0.10658698
#> 5       0.5 -0.795201560  1.17691425
#> 6       0.6  0.234326923  0.44739115
#> 7       0.7 -1.222451097  2.27295477
#> 8       0.8 -2.453647354  0.13605821
#> 9       0.9 -1.489260814 -1.99903913
#> 10      1.0 -0.432147734 -0.42050087
#> 11      1.1 -0.942554006 -0.37840740
#> 12      1.2 -0.121450799  1.22077479
#> 13      1.3  1.336446798 -1.54103029
#> 14      1.4 -0.860356182 -0.31031012
#> 15      1.5  0.666537820 -0.02010818
#> 16      1.6 -1.421534746 -2.39020034
#> 17      1.7  1.170056168  0.88986536
#> 18      1.8 -1.404714543 -1.48281332
#> 19      1.9  1.101708096  0.44575035
#> 20      2.0  0.697986263  1.36977586
#> 21      2.1 -0.864349803 -0.02011003
#> 22      2.2 -1.091470351 -0.10921759
#> 23      2.3 -0.037051465  0.26466174
#> 24      2.4  0.810053792  0.30384826
#> 25      2.5 -0.499355412 -0.18338848
#> 26      2.6  0.948031588  0.55964967
#> 27      2.7 -0.174245957 -0.18655384
#> 28      2.8 -1.106235952 -0.81227537
#> 29      2.9 -0.945985005 -1.64058167
#> 30      3.0  0.289089591  0.50792248
#> 31      3.1  0.876913145  1.75433696
#> 32      3.2 -1.148903940  0.59240020
#> 33      3.3 -1.137612756  1.01671329
#> 34      3.4 -1.437246735  0.12162059
#> 35      3.5 -0.494143476 -1.07806726
#> 36      3.6  0.840801808 -1.14356572
#> 37      3.7  0.791534124 -0.52964368
#> 38      3.8 -0.168848948 -0.68127316
#> 39      3.9  0.612722104 -0.20244756
#> 40      4.0 -0.771158924  1.68449572
#> 41      4.1  0.888628993 -1.03377324
#> 42      4.2  0.013214477 -0.15597667
#> 43      4.3  0.225339515 -0.04640064
#> 44      4.4 -0.729915210 -0.95362873
#> 45      4.5 -1.222487070  0.41626080
#> 46      4.6  0.406805171  0.11402961
#> 47      4.7 -0.751012223  0.06391875
#> 48      4.8 -0.162116540 -0.91933224
#> 49      4.9  0.352010126  0.90133529
#> 50      5.0 -0.289058300 -0.79772830
#> 51      0.1  0.104662227  0.66822120
#> 52      0.2  0.720186531  0.15521430
#> 53      0.3 -0.611046082  0.12868809
#> 54      0.4 -1.106914072 -1.53306545
#> 55      0.5  0.534803326  0.20236067
#> 56      0.6  0.736067968 -0.71753865
#> 57      0.7 -1.222501574  0.36169476
#> 58      0.8  1.021415310  1.39900429
#> 59      0.9  0.465165158  0.37269896
#> 60      1.0  0.790472705 -1.56564429
#> 61      1.1 -0.130264801 -0.05169454
#> 62      1.2 -0.930285334  0.51408210
#> 63      1.3 -0.364851004  0.54989952
#> 64      1.4  0.153872493  0.86781691
#> 65      1.5  0.413154818  0.68436008
#> 66      1.6  2.480823360 -0.16267998
#> 67      1.7 -2.179956742 -1.78436472
#> 68      1.8  0.420874578 -1.03714557
#> 69      1.9 -0.357528325  0.83014772
#> 70      2.0 -0.646861514  0.60734694
#> 71      2.1 -0.050141801 -0.12218636
#> 72      2.2  0.416942847  0.93312514
#> 73      2.3 -0.632587542 -0.96127668
#> 74      2.4  1.150146673  0.25508171
#> 75      2.5 -0.235475907 -0.54540157
#> 76      2.6 -1.643107386  0.93036074
#> 77      2.7 -1.503382146 -0.53765056
#> 78      2.8 -2.050584847 -0.45242607
#> 79      2.9 -0.753198229 -0.43929097
#> 80      3.0 -0.134141958 -0.61622311
#> 81      3.1  1.005782847  0.44163455
#> 82      3.2  2.167186798  0.48259745
#> 83      3.3  2.322556540  0.54214438
#> 84      3.4 -1.020423391 -2.29078465
#> 85      3.5  0.048814436  0.31035185
#> 86      3.6 -0.771888628  1.40407898
#> 87      3.7 -0.785235068  1.37711764
#> 88      3.8 -0.726603031  1.06039565
#> 89      3.9  0.681878032  0.63217289
#> 90      4.0 -0.229843287  1.08492804
#> 91      4.1 -1.510601724  1.35645944
#> 92      4.2 -0.583727687  0.36242404
#> 93      4.3 -2.022918454  2.16934446
#> 94      4.4  0.403504676  0.13913051
#> 95      4.5  0.550015549  1.37632653
#> 96      4.6  0.028357122 -0.49144999
#> 97      4.7  0.893165020  1.53491576
#> 98      4.8 -0.376555496 -0.41619872
#> 99      4.9  0.605884808 -0.52054380
#> 100     5.0 -0.004874726  0.85058387