Finalize data after analysis

Function that is used internally to change the data back to their original format after conducting an analysis, it being either binning ([bin()]) or using the Kalman filter ([kalman_filter()]). The finalization consists of renaming the columns to the user-defined column names instead of the package- required ones and deleting the grouping variable if it was not originally present in the data.

Usage

finalize(data, cols = NULL, .by = NULL)

Arguments

data

Dataframe that contains information on location (x- and y-coordinates) and the time at which the measurement was taken. By default, [kalman_filter()] will assume that this information is contained within the columns `"x"`, `"y"`, and `"time"` respectively.

cols

Named vector or named list containing the relevant column names in 'data' if they didn't contain the prespecified column names `"time"`, `"x"`, and `"y"`. The labels should conform to these prespecified column names and the values given to these locations should contain the corresponding column names in that dataset. Defaults to `NULL`, therefore assuming the structure explained in `data`.

.by

String denoting whether the moving window should be taken with respect to a given grouping variable. Defaults to `NULL`.

Value

Adjusted `data.frame`

Examples

# Generate data for illustration purposes
#
# Note that for this to work, I need to define the package-required column 
# names
data <- data.frame(
  x = rnorm(100),
  y = rnorm(100),
  time = rep(1:50, times = 2) / 10,
  id = rep(1:2, each = 50)
)

# Prepare the data for analysis
finalize(
  data,
  cols = c(
    "time" = "seconds",
    "x" = "X",
    "y" = "Y"
  ),
  .by = "tag"
)
#>     seconds            X           Y
#> 1       0.1  1.017491053  0.19795800
#> 2       0.2 -1.496053742 -1.20356611
#> 3       0.3 -1.184818730 -0.03981704
#> 4       0.4  0.630234373  0.68698246
#> 5       0.5  2.101252514  0.70526701
#> 6       0.6 -0.613736810  0.99144168
#> 7       0.7 -1.634638272  1.14424897
#> 8       0.8 -0.010441117 -1.23891024
#> 9       0.9 -0.656506139  2.65489833
#> 10      1.0 -0.669533441 -0.15691719
#> 11      1.1 -0.478589028 -0.42349012
#> 12      1.2  1.319456316 -0.19838706
#> 13      1.3  0.636562761 -0.89480241
#> 14      1.4  0.514327782  0.90426912
#> 15      1.5 -1.751375113  0.07964921
#> 16      1.6  0.893597518 -1.25882722
#> 17      1.7  0.223038372  1.02568511
#> 18      1.8  0.580816593 -0.73077860
#> 19      1.9 -0.177821421 -0.19014551
#> 20      2.0  0.740966708  0.52886469
#> 21      2.1 -0.997443079  0.55021053
#> 22      2.2 -2.938977561  0.54968434
#> 23      2.3  0.719015661 -0.65954237
#> 24      2.4 -0.698005041  0.05742171
#> 25      2.5 -1.894125843 -2.80801051
#> 26      2.6  0.076299249 -0.91225975
#> 27      2.7  0.875308501 -0.78237916
#> 28      2.8  0.453827393 -0.66410492
#> 29      2.9 -0.850716906  0.62630977
#> 30      3.0  0.566201613 -0.50724821
#> 31      3.1  1.152211954  0.27036133
#> 32      3.2 -0.756197377  0.46747686
#> 33      3.3 -0.489258334  0.72399496
#> 34      3.4 -1.166052337  0.61383694
#> 35      3.5 -0.479668950 -0.61786920
#> 36      3.6  0.115348218  0.22072490
#> 37      3.7 -1.768048407  1.12792660
#> 38      3.8 -1.407638919  1.81345434
#> 39      3.9  0.709178461 -0.08382568
#> 40      4.0 -1.240842940  1.36770667
#> 41      4.1 -0.368327348 -0.62743462
#> 42      4.2  0.462080093 -0.21662915
#> 43      4.3 -0.322833101 -0.68371382
#> 44      4.4 -1.287214810 -0.44470273
#> 45      4.5 -1.030040247  0.60648981
#> 46      4.6  1.514089316  0.62418307
#> 47      4.7  0.346903586 -0.69543107
#> 48      4.8  1.779441542 -0.78363908
#> 49      4.9  0.386630924 -0.95312386
#> 50      5.0 -0.918695239  1.79275607
#> 51      0.1 -1.584336488  0.34897670
#> 52      0.2 -0.084058892  0.25910377
#> 53      0.3 -2.085070889 -0.80595190
#> 54      0.4  0.003567992  0.10566470
#> 55      0.5 -0.355770822 -0.33359968
#> 56      0.6  1.146359751  1.64184797
#> 57      0.7 -0.221188446 -0.64390586
#> 58      0.8  1.018179021  0.58702056
#> 59      0.9 -0.263719295 -0.15040309
#> 60      1.0  1.658542305 -1.71082185
#> 61      1.1 -0.774086771  1.43103256
#> 62      1.2 -0.923937880 -2.64521227
#> 63      1.3 -0.275533378 -1.03245741
#> 64      1.4 -0.593399688 -0.70746643
#> 65      1.5 -0.122285891 -0.70056001
#> 66      1.6  1.179784246  0.53788544
#> 67      1.7  0.641037374 -0.31633217
#> 68      1.8 -0.629588508 -0.83962275
#> 69      1.9 -0.807734971 -1.35492806
#> 70      2.0 -0.860489929 -0.81756827
#> 71      2.1 -2.169238693 -0.63440000
#> 72      2.2 -1.375836518  0.81594943
#> 73      2.3 -0.493132472  0.30279571
#> 74      2.4 -0.581652027  1.80708662
#> 75      2.5 -0.167229304 -0.89402676
#> 76      2.6  0.485993129 -0.04642821
#> 77      2.7 -1.333395796 -0.47117914
#> 78      2.8 -0.261965625 -0.52669263
#> 79      2.9  0.652386303 -0.09513491
#> 80      3.0  0.748854971 -2.49536481
#> 81      3.1  0.896560285  0.16688922
#> 82      3.2  1.489300424  0.35049238
#> 83      3.3 -0.659403481  1.43370101
#> 84      3.4  0.537283179  0.76590680
#> 85      3.5  0.746803067  1.16752067
#> 86      3.6  1.896317084 -0.13694343
#> 87      3.7 -2.060070725 -0.51490204
#> 88      3.8  0.064543870  1.51974447
#> 89      3.9 -0.265147403 -0.32849168
#> 90      4.0 -0.447344531 -0.05367151
#> 91      4.1 -1.410700927 -0.56352463
#> 92      4.2 -0.506418882 -0.74390896
#> 93      4.3 -0.269761838 -0.10904165
#> 94      4.4 -1.085154918 -0.56082923
#> 95      4.5  0.362159127  0.18800155
#> 96      4.6 -0.335672143  0.74885094
#> 97      4.7  1.363804498 -1.91653832
#> 98      4.8 -0.711524136  0.23609585
#> 99      4.9  0.662178797  0.62895342
#> 100     5.0  0.291130223  0.41792568